refactor: move pages out of src
- adapt lint command
- bump deps
