+++ pages/api/auth/login.ts
@@ -1,12 +1,15 @@
 import nextConnect from 'next-connect';
 
-import { Provider } from '../../../client/context/AuthContext/AuthContext';
-import { encryptSession, setTokenCookie } from '../../../server/auth/cookie';
+import { Provider } from '../../../src/client/context/AuthContext/AuthContext';
+import {
+  encryptSession,
+  setTokenCookie,
+} from '../../../src/server/auth/cookie';
 import {
   promisifyAuthentication,
   authMiddleware,
-} from '../../../server/auth/middlewares';
-import { OK, UNAUTHORIZED, NOT_FOUND } from '../../../utils/statusCodes';
+} from '../../../src/server/auth/middlewares';
+import { OK, UNAUTHORIZED, NOT_FOUND } from '../../../src/utils/statusCodes';
 
 const provider: Provider = 'local';
