+++ pages/api/auth/logout.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 
-import { removeTokenCookie } from '../../../server/auth/cookie';
-import { FOUND_MOVED_TEMPORARILY } from '../../../utils/statusCodes';
+import { removeTokenCookie } from '../../../src/server/auth/cookie';
+import { FOUND_MOVED_TEMPORARILY } from '../../../src/utils/statusCodes';
 
 export default function handler(_: NextApiRequest, res: NextApiResponse) {
   removeTokenCookie(res);
