+++ pages/getting-started.tsx
