+++ pages/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import IndexPage from '../client/routes/Index';
+import IndexPage from '../src/client/routes/Index';
 
 export default function Index() {
   return <IndexPage />;
