+++ testUtils/index.tsx
@@ -3,15 +3,23 @@ import { render as rtlRender } from '@testing-library/react';
 import React, { cloneElement, ReactNode } from 'react';
 import { I18nextProvider, I18nContext } from 'react-i18next';
 
+import { AuthContextProvider } from '../src/client/context/AuthContext';
+import { User } from '../src/client/context/AuthContext/AuthContext';
 import i18n from './i18n';
 
 type TestOptions = {
   /**
    * whether to include the i18n.t prop
-   * @default true
    */
   includeTranslation?: boolean;
+  /**
+   * optional additional wrapper, e.g. Context
+   */
   wrapper?: typeof ChildrenPassthrough;
+  /**
+   * optional session to initialize AuthContextProvider with
+   */
+  session?: User | null;
 };
 
 function ChildrenPassthrough({ children }: { children: ReactNode }) {
@@ -23,6 +31,7 @@ function render(
   {
     includeTranslation = true,
     wrapper: AdditionalWrapper = ChildrenPassthrough,
+    session = null,
   }: TestOptions = {}
 ) {
   return {
@@ -30,17 +39,19 @@ function render(
       <ThemeProvider>
         <ColorModeProvider>
           <I18nextProvider i18n={i18n}>
-            <AdditionalWrapper>
-              <I18nContext.Consumer>
-                {({ t }) =>
-                  // this allows usage of render(<Component some="prop" />) in tests
-                  // translation won't break because the actual tFunction will be injected if necessary
-                  cloneElement(component, {
-                    t: includeTranslation ? t : undefined,
-                  })
-                }
-              </I18nContext.Consumer>
-            </AdditionalWrapper>
+            <AuthContextProvider session={session}>
+              <AdditionalWrapper>
+                <I18nContext.Consumer>
+                  {({ t }) =>
+                    // this allows usage of render(<Component some="prop" />) in tests
+                    // of components that receive `t` via props
+                    cloneElement(component, {
+                      t: includeTranslation ? t : undefined,
+                    })
+                  }
+                </I18nContext.Consumer>
+              </AdditionalWrapper>
+            </AuthContextProvider>
           </I18nextProvider>
         </ColorModeProvider>
       </ThemeProvider>
